添加了动态库的协处理器，优化了hbase2es模块的代码
