+++ Distribution/pom.xml
@@ -48,6 +48,11 @@
             <artifactId>streaming</artifactId>
             <version>1.0</version>
         </dependency>
+        <dependency>
+            <groupId>com.hzgc</groupId>
+            <artifactId>hbase2es</artifactId>
+            <version>1.0</version>
+        </dependency>
     </dependencies>
 
     <build>
