+++ HBase/src/main/java/com/hzgc/hbase/device/WarnRuleServiceImpl.java
@@ -6,8 +6,6 @@ import com.hzgc.hbase.util.HBaseHelper;
 import com.hzgc.hbase.util.HBaseUtil;
 import com.hzgc.util.ObjectUtil;
 import com.hzgc.util.StringUtil;
-import org.apache.hadoop.hbase.Cell;
-import org.apache.hadoop.hbase.CellScanner;
 import org.apache.hadoop.hbase.client.*;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.log4j.Logger;
