+++ HBase2Es/src/main/java/com/hzgc/hbase2es/es/ElasticSearchBulkOperator.java
@@ -1,0 +1,110 @@
+package com.hzgc.hbase2es.es;
+
+import com.hzgc.hbase2es.util.EsClientUtils;
+import org.elasticsearch.action.bulk.BulkRequestBuilder;
+import org.elasticsearch.action.bulk.BulkResponse;
+import org.elasticsearch.action.delete.DeleteRequestBuilder;
+import org.elasticsearch.action.support.WriteRequest;
+import org.elasticsearch.action.update.UpdateRequestBuilder;
+
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
+
+public class ElasticSearchBulkOperator {
+    private static final int MAX_BULK_COUNT = 10;
+    private static BulkRequestBuilder bulkRequestBuilder = null;
+
+    private static final Lock commitLock = new ReentrantLock();
+
+    private static ScheduledExecutorService scheduledExecutorService = null;
+
+    static {
+        // init es bulkRequestBuilder
+        bulkRequestBuilder = EsClientUtils.client.prepareBulk();
+        bulkRequestBuilder.setRefreshPolicy(WriteRequest.RefreshPolicy.IMMEDIATE);
+
+        // init thread pool and set size 1
+        scheduledExecutorService = Executors.newScheduledThreadPool(1);
+
+        // create beeper thread( it will be sync data to ES cluster)
+        // use a commitLock to protected bulk es as thread-save
+        final Runnable beeper = new Runnable() {
+            public void run() {
+                commitLock.lock();
+                try {
+                    bulkRequest(0);
+                } catch (Exception ex) {
+                    System.out.println(ex.getMessage());
+                } finally {
+                    commitLock.unlock();
+                }
+            }
+        };
+
+        // set time bulk task
+        // set beeper thread(10 second to delay first execution , 30 second period between successive executions)
+        scheduledExecutorService.scheduleAtFixedRate(beeper, 10, 30, TimeUnit.SECONDS);
+
+    }
+
+    /**
+     * shutdown time task immediately
+     */
+    public static void shutdownScheduEx() {
+        if (null != scheduledExecutorService && !scheduledExecutorService.isShutdown()) {
+            scheduledExecutorService.shutdown();
+        }
+    }
+
+    /**
+     * bulk request when number of builders is grate then threshold
+     *
+     * @param threshold
+     */
+    private static void bulkRequest(int threshold) {
+        if (bulkRequestBuilder.numberOfActions() > threshold) {
+            BulkResponse bulkItemResponse = bulkRequestBuilder.execute().actionGet();
+            if (!bulkItemResponse.hasFailures()) {
+                bulkRequestBuilder = EsClientUtils.client.prepareBulk();
+            }
+        }
+    }
+
+    /**
+     * add update builder to bulk
+     * use commitLock to protected bulk as thread-save
+     * @param builder
+     */
+    public static void addUpdateBuilderToBulk(UpdateRequestBuilder builder) {
+        commitLock.lock();
+        try {
+            bulkRequestBuilder.add(builder);
+            bulkRequest(MAX_BULK_COUNT);
+        } catch (Exception ex) {
+            ex.printStackTrace();
+        } finally {
+            commitLock.unlock();
+        }
+    }
+
+    /**
+     * add delete builder to bulk
+     * use commitLock to protected bulk as thread-save
+     *
+     * @param builder 删除索引的对象
+     */
+    public static void addDeleteBuilderToBulk(DeleteRequestBuilder builder) {
+        commitLock.lock();
+        try {
+            bulkRequestBuilder.add(builder);
+            bulkRequest(MAX_BULK_COUNT);
+        } catch (Exception ex) {
+            ex.printStackTrace();
+        } finally {
+            commitLock.unlock();
+        }
+    }
+}
