+++ HBase2Es/src/main/resources/es-person-index.properties
@@ -1,0 +1,5 @@
+es_cluster=my-cluster
+es_port=9300
+es_hosts=172.18.18.22
+es_index=dynamic
+es_type=person
