+++ HBase2Es/src/main/resources/es-static-index.properties
@@ -1,0 +1,5 @@
+es_cluster=my-cluster
+es_port=9300
+es_hosts=ldl
+es_index=objectinfo
+es_type=person
