--- HBase2Es/src/main/resources/es_cluster_config.properties
@@ -1,5 +1,0 @@
-es.cluster.name=my-cluster
-es.cluster.port=9300
-es.cluster.hosts=ldl
-es.cluster.index=objectinfo
-es.cluster.type=person
