+++ HBase2Es/src/main/resources/log4j.properties
@@ -1,19 +1,8 @@
-log4j.rootLogger=DEBUG,Console,Stdout
-
-#Console
-log4j.appender.Console=org.apache.log4j.ConsoleAppender
-log4j.appender.Console.layout=org.apache.log4j.PatternLayout
-log4j.appender.Console.layout.ConversionPattern=%d [%t] %-5p [%c] - %m%n
-
-log4j.logger.java.sql.ResultSet=INFO
-log4j.logger.org.apache=INFO
-log4j.logger.java.sql.Connection=DEBUG
-log4j.logger.java.sql.Statement=DEBUG
-log4j.logger.java.sql.PreparedStatement=DEBUG
-
-log4j.appender.Stdout = org.apache.log4j.DailyRollingFileAppender
-log4j.appender.Stdout.File = /home/HBase2EsObserver/log.txt
-log4j.appender.Stdout.Append = true
-log4j.appender.Stdout.Threshold = DEBUG
-log4j.appender.Stdout.layout = org.apache.log4j.PatternLayout
-log4j.appender.Stdout.layout.ConversionPattern = %-d{yyyy-MM-dd HH:mm:ss}  [ %t:%r ] - [ %p ]  %m%n
+log4j.rootLogger=INFO, R
+log4j.appender.R=org.apache.log4j.ConsoleAppender
+#log4j.appender.R=org.apache.log4j.RollingFileAppender
+#log4j.appender.R.File=log/ftpd.log
+#log4j.appender.R.MaxFileSize=100MB
+#log4j.appender.R.MaxBackupIndex=10
+log4j.appender.R.layout=org.apache.log4j.PatternLayout
+log4j.appender.R.layout.ConversionPattern=%d : %p : %c{1} : %m%n
