+++ HBase2Es/src/main/resources/person-table.properties
@@ -1,0 +1,6 @@
+table.name=person
+table.coprocessor.path=hdfs:///user/ldl/hbase2es-1.0-jar-with-dependencies.jar
+table.coprocessor.name=com.hzgc.hbase2es.coproccessor.PersonRepo2EsObserver
+table.colfams=i
+table.maxversion=1
+table.timetolive=15768000
