+++ HBase2Es/src/main/resources/static-table.properties
@@ -1,6 +1,6 @@
 table.name=objectinfo
 table.coprocessor.path=hdfs:///user/ldl/hbase2es-1.0-jar-with-dependencies.jar
-table.coprocessor.name=com.hzgc.hbase2es.HBase2EsObserver
+table.coprocessor.name=com.hzgc.hbase2es.coproccessor.StaticRepo2EsObserver
 table.colfams=person
 table.maxversion=1
 table.timetolive=15768000
